requirements.txt
# ----------------------
# Core Flask stack
# ----------------------
Flask==3.1.1
click==8.2.1
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
blinker==1.9.0

# ----------------------
# Image processing / Math
# ----------------------
pillow==11.2.1
numpy==2.3.0
mpmath==1.3.0
sympy==1.14.0

# ----------------------
# File handling and utils
# ----------------------
filelock==3.18.0
fsspec==2025.5.1
typing_extensions==4.14.0
networkx==3.5

# ----------------------
# Torch stack with CUDA 12.1 (NVIDIA T4 compatible)
# ----------------------
torch==2.2.2+cu121
torchvision==0.17.2+cu121
torchaudio==2.2.2+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

# ----------------------
# Hugging Face stack for pretrained transformer & diffusion models
# ----------------------
transformers==4.41.1
diffusers==0.28.0
accelerate==0.30.0
